Move end BP to its own folder
