+++ src/main/java/voronoiaoc/byg/common/world/dimension/end/BYGEndLayerProvider.java
@@ -1,4 +1,4 @@
-package voronoiaoc.byg.common.world.dimension;
+package voronoiaoc.byg.common.world.dimension.end;
 
 
 import net.minecraft.world.biome.layer.ScaleLayer;
